+++ src/main/java/net/neoforged/neoforge/transfer/fluid/base/EmptyFluidStorage.java
@@ -1,6 +1,7 @@
-package net.neoforged.neoforge.transfer.fluid;
+package net.neoforged.neoforge.transfer.fluid.base;
 
-import net.neoforged.neoforge.transfer.storage.EmptyStorage;
+import net.neoforged.neoforge.transfer.fluid.FluidVariant;
+import net.neoforged.neoforge.transfer.storage.base.EmptyStorage;
 
 public final class EmptyFluidStorage extends EmptyStorage<FluidVariant> {
     public static final EmptyFluidStorage INSTANCE = new EmptyFluidStorage();
