+++ src/main/java/net/neoforged/neoforge/transfer/fluid/base/FluidHandlerAdapter.java
@@ -3,10 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.neoforge.transfer.fluid;
+package net.neoforged.neoforge.transfer.fluid.base;
 
 import net.neoforged.neoforge.fluids.FluidStack;
 import net.neoforged.neoforge.fluids.capability.IFluidHandler;
+import net.neoforged.neoforge.transfer.fluid.FluidVariant;
 import net.neoforged.neoforge.transfer.storage.Storage;
 import net.neoforged.neoforge.transfer.storage.StorageUtil;
 import net.neoforged.neoforge.transfer.transaction.Transaction;
