+++ src/main/java/net/neoforged/neoforge/transfer/item/base/EmptyItemStorage.java
@@ -1,6 +1,7 @@
-package net.neoforged.neoforge.transfer.item;
+package net.neoforged.neoforge.transfer.item.base;
 
-import net.neoforged.neoforge.transfer.storage.EmptyStorage;
+import net.neoforged.neoforge.transfer.item.ItemVariant;
+import net.neoforged.neoforge.transfer.storage.base.EmptyStorage;
 
 public final class EmptyItemStorage extends EmptyStorage<ItemVariant> {
     public static final EmptyItemStorage INSTANCE = new EmptyItemStorage();
