+++ src/main/java/net/neoforged/neoforge/transfer/storage/base/CombinedStorage.java
@@ -1,5 +1,7 @@
-package net.neoforged.neoforge.transfer.storage;
+package net.neoforged.neoforge.transfer.storage.base;
 
+import net.neoforged.neoforge.transfer.storage.Storage;
+import net.neoforged.neoforge.transfer.storage.StoragePreconditions;
 import net.neoforged.neoforge.transfer.transaction.TransactionContext;
 
 /**
